requirements.txt
torch
matplotlib
pandas
numpy
sklearn
Pillow